Modify Calculator.java divide method

+++ src/week3_day2/Calculator.java
@@ -11,6 +11,6 @@ public class Calculator {
         System.out.println(a*b);
     }
     public void divide(int a, int b){
-        System.out.println(a/b);
+        System.out.println((double) a/b);
     }
 }
